+++ Flink/FraudDetection/src/main/java/FraudDetection/Output_Event.java
@@ -1,8 +1,8 @@
 /**************************************************************************************
  *  Copyright (c) 2019- Gabriele Mencagli and Alessandra Fais
- *  
+ *
  *  This file is part of StreamBenchmarks.
- *  
+ *
  *  StreamBenchmarks is free software dual licensed under the GNU LGPL or MIT License.
  *  You can redistribute it and/or modify it under the terms of the
  *    * GNU Lesser General Public License as published by
@@ -10,7 +10,7 @@
  *      (at your option) any later version
  *    OR
  *    * MIT License: https://github.com/ParaGroup/StreamBenchmarks/blob/master/LICENSE.MIT
- *  
+ *
  *  StreamBenchmarks is distributed in the hope that it will be useful,
  *  but WITHOUT ANY WARRANTY; without even the implied warranty of
  *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
@@ -25,16 +25,16 @@ package FraudDetection;
 
 public class Output_Event {
 	public String entityID;
-	public double score;
+	public float score;
 	public long ts;
 
 	public Output_Event() {
 		entityID = "";
-		score = 0d;
+		score = 0.0f;
 		ts = 0L;
 	}
 
-	public Output_Event(String _entityID, double _score, long _ts) {
+	public Output_Event(String _entityID, float _score, long _ts) {
 		entityID = _entityID;
 		score = _score;
 		ts = _ts;
