+++ Flink/FraudDetection/src/main/java/MarkovModelPrediction/Prediction.java
@@ -1,8 +1,8 @@
 /**************************************************************************************
  *  Copyright (c) 2019- Gabriele Mencagli and Alessandra Fais
- *  
+ *
  *  This file is part of StreamBenchmarks.
- *  
+ *
  *  StreamBenchmarks is free software dual licensed under the GNU LGPL or MIT License.
  *  You can redistribute it and/or modify it under the terms of the
  *    * GNU Lesser General Public License as published by
@@ -10,7 +10,7 @@
  *      (at your option) any later version
  *    OR
  *    * MIT License: https://github.com/ParaGroup/StreamBenchmarks/blob/master/LICENSE.MIT
- *  
+ *
  *  StreamBenchmarks is distributed in the hope that it will be useful,
  *  but WITHOUT ANY WARRANTY; without even the implied warranty of
  *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
@@ -31,11 +31,11 @@ import java.io.Serializable;
  */
 public class Prediction implements Serializable {
     private String entityId;
-    private double score;
+    private float score;
     private String[] states;
     private boolean outlier;
 
-    public Prediction(String entityId, double score, String[] states, boolean outlier) {
+    public Prediction(String entityId, float score, String[] states, boolean outlier) {
         this.entityId = entityId;
         this.score = score;
         this.states = states;
@@ -50,11 +50,11 @@ public class Prediction implements Serializable {
         this.entityId = entityId;
     }
 
-    public double getScore() {
+    public float getScore() {
         return score;
     }
 
-    public void setScore(double score) {
+    public void setScore(float score) {
         this.score = score;
     }
